fix(DetailedPresence): fetch all calls
PubNub Presence notification doesn't return all active calls,this behavior is by design. Because of
payload size limits we provide up to 7 active calls in Presence Push notifications. So we need to
verify `totalActiveCalls` parameter and if the total count is larger than the actual list length,
then we need to go to the Presence API to get all the details.

https://jira.ringcentral.com/browse/PLA-33391